cast to double before tiledist calc

+++ src/binding.cpp
@@ -1023,8 +1023,8 @@ inline bool contextSortByRelev(Context const& a, Context const& b) noexcept {
 inline double tileDist(unsigned px, unsigned py, unsigned tileX, unsigned tileY) {
     const unsigned tileCenterX = tileX;
     const unsigned tileCenterY = tileY;
-    const unsigned dx = px - tileCenterX;
-    const unsigned dy = py - tileCenterY;
+    const double dx = static_cast<double>(px - tileCenterX);
+    const double dy = static_cast<double>(py - tileCenterY);
     const double distance = dx * dx + dy * dy;
 
     return distance;
